# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectorInstallation": {
        "additionalProperties": false,
        "description": "不暴露配置引用的 Connector 管理投影。",
        "properties": {
          "adapterKey": {
            "title": "Adapterkey",
            "type": "string"
          },
          "allowedOrigins": {
            "items": {
              "type": "string"
            },
            "title": "Allowedorigins",
            "type": "array"
          },
          "configurationState": {
            "const": "CONFIGURED",
            "default": "CONFIGURED",
            "title": "Configurationstate",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "environmentId": {
            "format": "uuid",
            "title": "Environmentid",
            "type": "string"
          },
          "healthState": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderHealthState"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "implementationVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implementationversion"
          },
          "installationKey": {
            "title": "Installationkey",
            "type": "string"
          },
          "lastValidatedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastvalidatedat"
          },
          "mode": {
            "$ref": "#/components/schemas/ConnectorMode"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "negotiatedCapabilities": {
            "items": {
              "$ref": "#/components/schemas/CapabilityDescriptor"
            },
            "title": "Negotiatedcapabilities",
            "type": "array"
          },
          "projectId": {
            "format": "uuid",
            "title": "Projectid",
            "type": "string"
          },
          "protocolVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocolversion"
          },
          "requiredCapabilities": {
            "items": {
              "$ref": "#/components/schemas/ProviderCapability"
            },
            "title": "Requiredcapabilities",
            "type": "array"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "safeMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safemessage"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectorStatus"
          },
          "tenantId": {
            "format": "uuid",
            "title": "Tenantid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "projectId",
          "environmentId",
          "installationKey",
          "name",
          "adapterKey",
          "mode",
          "allowedOrigins",
          "requiredCapabilities",
          "negotiatedCapabilities",
          "status",
          "healthState",
          "safeMessage",
          "protocolVersion",
          "implementationVersion",
          "lastValidatedAt",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "title": "ConnectorInstallation",
        "type": "object"
      },
      "ConnectorInstallationPage": {
        "additionalProperties": false,
        "description": "ConnectorInstallation Cursor Page。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConnectorInstallation"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ConnectorInstallationPage",
        "type": "object"
      },
...
      "CreateConnectorInstallation": {
        "additionalProperties": false,
        "description": "在一个 Environment 中登记显式 Adapter 与安全配置引用。",
        "properties": {
          "adapterKey": {
            "pattern": "^[a-z][a-z0-9-]{1,63}$",
            "title": "Adapterkey",
            "type": "string"
          },
          "allowedOrigins": {
            "items": {
              "type": "string"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Allowedorigins",
            "type": "array"
          },
          "configurationRef": {
            "maxLength": 204,
            "minLength": 12,
            "pattern": "^cfg_[A-Za-z0-9_-]{8,200}$",
            "title": "Configurationref",
            "type": "string"
          },
          "environmentId": {
            "format": "uuid",
            "title": "Environmentid",
            "type": "string"
          },
          "installationKey": {
            "maxLength": 64,
            "minLength": 2,
            "pattern": "^[a-z][a-z0-9._-]{1,63}$",
            "title": "Installationkey",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ConnectorMode"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "requiredCapabilities": {
            "items": {
              "$ref": "#/components/schemas/ProviderCapability"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Requiredcapabilities",
            "type": "array"
          }
        },
        "required": [
          "environmentId",
          "installationKey",
          "name",
          "adapterKey",
          "mode",
          "configurationRef",
          "allowedOrigins",
          "requiredCapabilities"
        ],
        "title": "CreateConnectorInstallation",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Atlas 扩展的 RFC Problem Details。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errorCode": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "errorCode",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Atlas TestOps Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/connector-installations": {
      "post": {
        "description": "登记安全配置引用；公共响应不回显 configurationRef。",
        "operationId": "create_connector_installation_v1_connector_installations_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectorInstallation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorInstallation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Connector 策略或请求无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "状态、唯一键或幂等冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "可信 Adapter 未安装"
          }
        },
        "summary": "创建 ConnectorInstallation",
        "tags": [
          "connectors"
        ]
      }
    },
    "/v1/connector-installations/{connectorId}": {
      "get": {
        "operationId": "get_connector_installation_v1_connector_installations__connectorId__get",
        "parameters": [
          {
            "in": "path",
            "name": "connectorId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connectorid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorInstallation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Connector 策略或请求无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "状态、唯一键或幂等冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "可信 Adapter 未安装"
          }
        },
        "summary": "读取 ConnectorInstallation",
        "tags": [
          "connectors"
        ]
      },
      "patch": {
        "operationId": "update_connector_installation_v1_connector_installations__connectorId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "connectorId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connectorid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConnectorInstallation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorInstallation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Connector 策略或请求无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "状态、唯一键或幂等冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "可信 Adapter 未安装"
          }
        },
        "summary": "更新 ConnectorInstallation",
        "tags": [
          "connectors"
        ]
      }
    },
    "/v1/connector-installations/{connectorId}:validate": {
      "post": {
        "description": "事务外执行 Probe，并以请求携带的 Revision CAS 写入结果。",
        "operationId": "validate_connector_installation_v1_connector_installations__connectorId__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "connectorId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connectorid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "maxLength": 64,
              "title": "If-Match",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorInstallation"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Connector 策略或请求无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "状态、唯一键或幂等冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "可信 Adapter 未安装"
          }
        },
        "summary": "验证 Connector 并刷新 Capability Snapshot",
        "tags": [
          "connectors"
        ]
      }
    },
...
    "/v1/environments/{environmentId}/connector-installations": {
      "get": {
        "description": "按稳定 Cursor 列出 Environment 的 Connector。",
        "operationId": "list_connector_installations_v1_environments__environmentId__connector_installations_get",
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Environmentid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorInstallationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Connector 策略或请求无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少有效身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "当前 PlatformRole 无权执行"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "状态、唯一键或幂等冲突"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Revision 前置条件失败"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "请求不符合接口契约"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "可信 Adapter 未安装"
          }
        },
        "summary": "列出 ConnectorInstallation",
        "tags": [
          "connectors"
        ]
      }
    },
...
  }
}